% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualizar-dados.R
\name{atualizar_dados}
\alias{atualizar_dados}
\title{Atualizar base de dados do Tidytuesday}
\usage{
atualizar_dados(file, ...)
}
\arguments{
\item{file}{Caminho para salvar a base (deve ser CSV)}

\item{...}{Outros arguementos passados para \code{\link[readr:write_delim]{readr::write_csv()}}}
}
\value{
Invisivel, a base atualizada
}
\description{
Ler dados da Tidytuesday de 20/04/2021, salvando em um arquivo CSV conforme
o argumento \code{file}. Para mais informações, acesse
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2021/2021-04-20}.
}
