% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{netflix}
\alias{netflix}
\title{Base Netflix do Tidytuesday}
\format{
O formato da base:
\describe{
\item{coluna}{descricao}
}
}
\source{
\url{https://www.kaggle.com/shivamb/netflix-shows?select=netflix_titles.csv}
}
\usage{
netflix
}
\description{
Uma base com informacoes sobre o Netflix utilizada no Tidytuesday de
20/04/2021.
}
\keyword{datasets}
