% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualizar-dados.R
\name{atualizar_dados}
\alias{atualizar_dados}
\title{Atualizar base de dados do Tidytuesday}
\usage{
atualizar_dados(file, ...)
}
\arguments{
\item{file}{caminho para salvar a base (deve ser CSV)}

\item{...}{Outros argumentos passados para \code{\link[readr:write_delim]{readr::write_csv()}}}
}
\value{
Invisivel, a base atualizada
}
\description{
Ler dados da Tidytuesday de 20/04/2021, salvando em um arquivo CSV conforme
o argumento \code{file}.
}
